refactor: divide Computer class
